Commit -update do get default lists, router, dynamic port
- Added get function for get default lists
- Added header to csv
- Added .env file to get port number
