+++ public/mockdb/defaultLists.csv
@@ -1,3 +1,4 @@
+name,file
 Home,home
-Upcoming, upcoming
-Filters, filters
+Upcoming,upcoming
+Filters,filters
